mainboard/winnet/g170: Copy from mainboard/bcom/winnetp680
G170 is a board manufactured by WinNET, used in thin clients including
HP Neoware CA19 and IGEL 2110.

Copied from mainboard/bcom/winnetp680 which seems to be a similar
board with an extra PCI slot.

The p680 should probably be moved to winnet/ too, since the board is an OEM
WinNET board, with BCom being just a machine that happens to it.

Change-Id: I90b89ee634d90cfba2e56cca5b76cfd2bd7a8d0b
Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>